Offset walk animations by 1 frame
Previously the first frame of the walking animations was the same as the idle texture. With the recent change making it so that the animation always restarts when an entity begins moving, this meant that it looked as though the animation didn't start playing right away when you started moving. If you moved back and forth quickly it would look like you were just staying as the idle sprite. Now the first frame of the walk animations are different from the idle sprite.