Bandaid fix for bug with entities spawning inside other entities.
When an entity is colliding with something before moving it will now just not move. Previously it would just teleport in the opposite direction of its movement to the nearest available space. Had to also fix the player being able to spawn inside obstacles at the start of the game otherwise the player wouldn't be able to move.